requirements.txt
fastapi
uvicorn
python-dotenv
pydantic[email]
SQLAlchemy
alembic
pymysql
passlib[bcrypt]
python-jose[cryptography]
python-multipart
requests
httpx
python-slugify
ctransformers
pydantic-settings
cryptography


